Offloading more component type definitions
