+++ src/frontend/src/tables/RowActions.tsx
@@ -1,3 +1,7 @@
+import { cancelEvent } from '@lib/functions/Events';
+import { getDetailUrl } from '@lib/functions/Navigation';
+import { navigateToLink } from '@lib/functions/Navigation';
+import type { RowAction, RowViewProps } from '@lib/types/Tables';
 import { t } from '@lingui/core/macro';
 import { ActionIcon, Menu, Tooltip } from '@mantine/core';
 import {
@@ -9,31 +13,8 @@ import {
   IconTrash
 } from '@tabler/icons-react';
 import { type ReactNode, useMemo, useState } from 'react';
-import type { NavigateFunction } from 'react-router-dom';
 
-import type { ModelType } from '@lib/enums/ModelType';
-import { cancelEvent } from '@lib/functions/Events';
-import { getDetailUrl } from '@lib/functions/Navigation';
-import { navigateToLink } from '@lib/functions/Navigation';
-
-// Type definition for a table row action
-export type RowAction = {
-  title?: string;
-  tooltip?: string;
-  color?: string;
-  icon?: ReactNode;
-  onClick?: (event: any) => void;
-  hidden?: boolean;
-  disabled?: boolean;
-};
-
-type RowModelProps = {
-  modelType: ModelType;
-  modelId: number;
-  navigate: NavigateFunction;
-};
-
-export type RowViewProps = RowAction & RowModelProps;
+export type { RowAction, RowViewProps } from '@lib/types/Tables';
 
 // Component for viewing a row in a table
 export function RowViewAction(props: RowViewProps): RowAction {
